feat: Update ShopListView to use ShopService for retrieving shops
This commit updates the ShopListView in the shop_views.py file to use
the ShopService for retrieving shops. It injects the ShopService
dependency into the view and modifies the get_queryset method to call
the get_all_shops method of the ShopService. This change improves the
separation of concerns and makes the code more modular and testable.