feat: Add create shop form and template
This commit adds a new form and template for creating a shop. The
`CreateShopForm` class is created in the `shop_forms.py` file, which
extends the `ModelForm` class and specifies the `Shop` model and the
fields to include in the form. The `shop_create.html` template is also
added, which extends the `extend_base.html` template and includes the
form for creating a shop. These changes allow users to create new shops
with a title and description.