feat: Add get_absolute_url method to Shop model
This commit adds a new method, get_absolute_url, to the Shop model in
the core/models.py file. The get_absolute_url method generates the URL
for the detail view of a shop using the reverse function. This addition
enhances the functionality of the Shop model by providing a convenient
way to retrieve the URL for a specific shop.